Refactor: "둘로 나눠진 Recoil Store 병합"
