+++ package.json
@@ -45,7 +45,6 @@
     ]
   },
   "devDependencies": {
-    "eslint": "^8.2.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-import": "^2.25.3",
     "eslint-plugin-jsx-a11y": "^6.5.1",
