--- src/state/StateBillList.js
@@ -1,18 +1,0 @@
-import {atom} from "recoil";
-
-export const billListState = atom({
-  key: "billList",
-  default: [],
-});
-export const pageState = atom({
-  key: "page",
-  default: 1,
-});
-export const excelFilterState = atom({
-  key: "excelFilter",
-  default: "전체",
-});
-export const billSizeState = atom({
-  key: "billSize",
-  default: 0,
-});
